Move the AntiForgeryToken file to client/client_modules folder.
- The client folder is shipped to browser.
- The client_modules folder is discovered by caller from `require`.
- Update the CSPROJ reference.
- Update webpack resolve path reference.
