+++ Nancy.ViewEngines.React/Nancy.ViewEngines.React.csproj
@@ -85,6 +85,7 @@
     </None>
     <None Include="targets\Nancy.ViewEngines.React.targets" />
     <None Include="tools\build.js" />
+    <None Include="tools\client\client_modules\anti-forgery-token.jsx" />
     <None Include="tools\client\console.js" />
     <None Include="tools\client\html.jsx" />
     <None Include="tools\client\invokeOrDefault.js" />
@@ -98,7 +99,6 @@
     <None Include="tools\install.ps1" />
     <None Include="tools\node.cmd" />
     <None Include="tools\uninstall.ps1" />
-    <None Include="tools\utils\anti-forgery-token.jsx" />
   </ItemGroup>
   <ItemGroup>
     <Analyzer Include="..\packages\StyleCop.Analyzers.1.0.0\analyzers\dotnet\cs\Newtonsoft.Json.dll" />
