+++ Nancy.ViewEngines.React/tools/client/client_modules/anti-forgery-token.jsx
